Delete main.yml
Because I've realized that this should always be explicitly defined in host_vars foreach host. If it is left undefined, I want it to fail as safe as possible; so, not only does it never assume the machine's a dev machine (and thus accidentally open holes in the security policy which would be unacceptable in prod) but it doesn't assume anything about prod status, and if that matters and isn't defined, the playbook fails.