switched openvpn topology config
The default is net30; I uncommented the line that overrides the default and makes it subnet. Why? Well, for one, the openvpn team actually recommends subnet over net30, and mentions that the only good reason to use net30 is if you must support an earlier version of openvpn running on Windows clients. Which I sure as hell won't be doing. Also, it necessitates each client getting four addresses, and allocating four addresses to the server (!). What the hell is that fresh nonsense? So, subnet it is.